refactor: create new service packages
This allows faster navigation inside of the project.
