+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:lifting_progress_tracker/app.dart';
-import 'package:lifting_progress_tracker/services/auth_service.dart';
-import 'package:lifting_progress_tracker/services/firebase_service.dart';
+import 'package:lifting_progress_tracker/services/firebase/auth_service.dart';
+import 'package:lifting_progress_tracker/services/firebase/firebase_service.dart';
 import 'package:logging/logging.dart';
 
 void main() {
