+++ lib/services/firebase/firebase_service.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:get_it/get_it.dart';
 import 'package:lifting_progress_tracker/firebase_options.dart';
-import 'package:lifting_progress_tracker/services/firestore_service.dart';
+import 'package:lifting_progress_tracker/services/firebase/firestore_service.dart';
 import 'package:logging/logging.dart';
 
 class FirebaseService {
